Vocab := Finset (Sum Nat Nat)
Note that both atomic propositions and atomic programs are `Nat``.
That is still the case now. But a `Vocab` now keeps track of
whether a `Nat`` was used as one or as the other kind of atom.

Previously a vocabulary was just a `Finset Nat` such that for example
the two PDL formulas `[1]2` and `[2]1` had the same vocabulary {1,2}
which affects the meaning of interpolation in a weird way.